requirements.txt
click
pynacl
subnet-utils
